refactor: moved block import utils to a common place to use
