+++ crates/client/block_import/src/tests.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod block_import_utils;
