+++ crates/client/block_import/src/verify_apply.rs
@@ -298,151 +298,25 @@ fn block_hash(
 #[cfg(test)]
 mod verify_apply_tests {
     use super::*;
-    use mc_db::tests::test_block::{finalized_block_zero, finalized_state_diff_zero};
-    use mp_block::header::{GasPrices, L1DataAvailabilityMode};
-    use mp_block::Header;
-    use mp_chain_config::{ChainConfig, StarknetVersion};
-    use mp_receipt::{
-        DeclareTransactionReceipt, DeployAccountTransactionReceipt, DeployTransactionReceipt, InvokeTransactionReceipt,
-        L1HandlerTransactionReceipt,
-    };
+    use crate::tests::block_import_utils::*;
+    use mc_db::tests::common::{finalized_block_zero, finalized_state_diff_zero};
+
+    use mp_chain_config::ChainConfig;
+
     use mp_state_update::{ContractStorageDiffItem, DeployedContractItem, StateDiff, StorageEntry};
-    use mp_transactions::{
-        DeclareTransactionV1, DeployAccountTransactionV3, DeployTransaction, InvokeTransactionV1, L1HandlerTransaction,
-    };
+
     use mp_utils::tests_common::set_workdir;
     use rstest::*;
     use starknet_api::{core::ChainId, felt};
     use std::sync::Arc;
 
-    /// Creates a dummy UnverifiedHeader for testing purposes.
-    ///
-    /// This function generates an UnverifiedHeader with predefined values,
-    /// useful for creating consistent test scenarios.
-    fn create_dummy_unverified_header() -> UnverifiedHeader {
-        UnverifiedHeader {
-            parent_block_hash: Some(felt!("0x1")),
-            sequencer_address: felt!("0x2"),
-            block_timestamp: 12345,
-            protocol_version: StarknetVersion::new(0, 13, 2, 0),
-            l1_gas_price: GasPrices {
-                eth_l1_gas_price: 14,
-                strk_l1_gas_price: 15,
-                eth_l1_data_gas_price: 16,
-                strk_l1_data_gas_price: 17,
-            },
-            l1_da_mode: L1DataAvailabilityMode::Blob,
-        }
-    }
-
-    /// Creates dummy ValidatedCommitments for testing purposes.
-    ///
-    /// This function generates ValidatedCommitments with predefined values,
-    /// useful for creating consistent test scenarios.
-    fn create_dummy_commitments() -> ValidatedCommitments {
-        ValidatedCommitments {
-            transaction_count: 1,
-            transaction_commitment: felt!("0x6"),
-            event_count: 2,
-            event_commitment: felt!("0x7"),
-            state_diff_length: 3,
-            state_diff_commitment: felt!("0x8"),
-            receipt_commitment: felt!("0x9"),
-        }
-    }
-
     /// Sets up a test backend.
     ///
-    /// This function creates a new MadaraBackend instance with a temporary
-    /// directory and test configuration, useful for isolated test environments.
+    /// This function creates a new MadaraBackend instance with a test configuration, useful for isolated test environments.
     #[fixture]
     pub fn setup_test_backend(_set_workdir: ()) -> Arc<MadaraBackend> {
         let chain_config = Arc::new(ChainConfig::test_config().unwrap());
         MadaraBackend::open_for_testing(chain_config.clone())
-    }
-
-    /// Creates a BlockValidationContext for testing purposes.
-    ///
-    /// This function generates a BlockValidationContext with customizable
-    /// ignore_block_order flag, useful for testing different validation scenarios.
-    fn create_validation_context(ignore_block_order: bool) -> BlockValidationContext {
-        BlockValidationContext {
-            chain_id: ChainId::Other("something".to_string()),
-            ignore_block_order,
-            trust_global_tries: false,
-            trust_transaction_hashes: false,
-            trust_class_hashes: false,
-        }
-    }
-
-    /// Creates a dummy Header for testing purposes.
-    ///
-    /// This function generates a Header with predefined values,
-    /// useful for creating consistent test scenarios.
-    fn create_dummy_header() -> Header {
-        Header {
-            parent_block_hash: felt!("0x1"),
-            sequencer_address: felt!("0x2"),
-            block_number: 1,
-            global_state_root: felt!("0xa"),
-            transaction_count: 0,
-            transaction_commitment: Felt::ZERO,
-            event_count: 0,
-            event_commitment: Felt::ZERO,
-            state_diff_length: 0,
-            state_diff_commitment: Felt::ZERO,
-            receipt_commitment: Felt::ZERO,
-            block_timestamp: 12345,
-            protocol_version: StarknetVersion::new(0, 13, 2, 0),
-            l1_gas_price: GasPrices {
-                eth_l1_gas_price: 14,
-                strk_l1_gas_price: 15,
-                eth_l1_data_gas_price: 16,
-                strk_l1_data_gas_price: 17,
-            },
-            l1_da_mode: L1DataAvailabilityMode::Blob,
-        }
-    }
-
-    /// Creates a dummy PreValidatedBlock for testing purposes.
-    ///
-    /// This function generates a PreValidatedBlock with predefined values,
-    /// useful for testing update_tries scenarios.
-    fn create_dummy_block() -> PreValidatedBlock {
-        PreValidatedBlock {
-            header: create_dummy_unverified_header(),
-            unverified_block_hash: None,
-            unverified_block_number: Some(1),
-            unverified_global_state_root: Some(felt!("0xa")),
-            commitments: create_dummy_commitments(),
-            transactions: vec![],
-            receipts: vec![],
-            state_diff: StateDiff::default(),
-            converted_classes: Default::default(),
-        }
-    }
-
-    /// Creates a dummy PreValidatedPendingBlock for testing purposes.
-    fn create_dummy_pending_block() -> PreValidatedPendingBlock {
-        PreValidatedPendingBlock {
-            header: UnverifiedHeader {
-                parent_block_hash: Some(felt!("0x1")),
-                sequencer_address: felt!("0x2"),
-                block_timestamp: 12345,
-                protocol_version: StarknetVersion::new(0, 13, 2, 0),
-                l1_gas_price: GasPrices {
-                    eth_l1_gas_price: 14,
-                    strk_l1_gas_price: 15,
-                    eth_l1_data_gas_price: 16,
-                    strk_l1_data_gas_price: 17,
-                },
-                l1_da_mode: L1DataAvailabilityMode::Blob,
-            },
-            transactions: vec![],
-            receipts: vec![],
-            state_diff: StateDiff::default(),
-            converted_classes: Default::default(),
-        }
     }
 
     /// Test various scenarios for the `check_parent_hash_and_num` function.
@@ -750,10 +624,10 @@ mod verify_apply_tests {
     mod verify_apply_inner_tests {
         use super::*;
 
-        /// Test successful block verification and storage.
+        /// Test successful pending block verification and storage.
         ///
         /// Verifies that:
-        /// 1. The function correctly verifies and stores a new block.
+        /// 1. The function correctly verifies and stores a new pending block.
         /// 2. The latest block number is updated after successful storage.
         #[rstest]
         #[tokio::test]
@@ -771,12 +645,12 @@ mod verify_apply_tests {
             block.unverified_global_state_root = Some(felt!("0x0"));
             let validation = create_validation_context(false);
 
-            let _result = verify_apply_inner(&backend, block.clone(), validation.clone());
+            let _result = verify_apply_inner(&backend, block, validation.clone());
 
             assert_eq!(backend.get_latest_block_n().unwrap(), Some(1));
         }
 
-        /// Test error handling during block verification.
+        /// Test error handling during pending block verification.
         ///
         /// Verifies that:
         /// 1. The function returns an error for invalid block data (e.g., mismatched block number).
@@ -798,7 +672,7 @@ mod verify_apply_tests {
 
             let validation = create_validation_context(false);
 
-            let result = verify_apply_inner(&backend, block.clone(), validation);
+            let result = verify_apply_inner(&backend, block, validation);
 
             assert!(matches!(result.unwrap_err(), BlockImportError::LatestBlockN { .. }));
             assert_eq!(backend.get_latest_block_n().unwrap(), Some(0));
