+++ crates/client/db/src/tests.rs
@@ -1,3 +1,4 @@
 pub mod common;
 pub mod test_block;
+#[cfg(test)]
 pub mod test_open;
